requirements.txt
filetype==1.2.0
# PySimpleGUI==5.0.8.3
pandas==2.2.3
openpyxl==3.1.5
rsa==4.9
numpy==2.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.2
et-xmlfile==2.0.0
six==1.17.0
pyasn1==0.6.1
polars==0.19.12
flask==2.3.2
flaskwebgui==1.1.6
pyinstaller==6.14.1
waitress==3.0.2
pywebview==5.4
pytest-mock==3.14.1